Bring `parse_blazemeter_test_link` to blazemeter.py
