+++ bzt/modules/blazemeter.py
@@ -24,10 +24,11 @@ import time
 import traceback
 import zipfile
 from abc import abstractmethod
-from collections import defaultdict, OrderedDict, Counter
+from collections import defaultdict, OrderedDict, Counter, namedtuple
 from functools import wraps
 from ssl import SSLError
 
+import re
 import yaml
 from requests.exceptions import ReadTimeout
 from urwid import Pile, Text
@@ -44,7 +45,7 @@ from bzt.modules.monitoring import Monitoring, MonitoringListener
 from bzt.modules.services import Unpacker
 from bzt.six import BytesIO, iteritems, HTTPError, r_input, URLError, b, string_types, text_type
 from bzt.utils import open_browser, get_full_path, get_files_recursive, replace_in_config, humanize_bytes, \
-    ExceptionalDownloader, ProgressBarContext, parse_blazemeter_test_link
+    ExceptionalDownloader, ProgressBarContext
 from bzt.utils import to_json, dehumanize_time, BetterDict, ensure_is_dict
 
 TAURUS_TEST_TYPE = "taurus"
@@ -143,6 +144,25 @@ def send_with_retry(method):
                 self.log.warning("Will skip failed data and continue running")
 
     return _impl
+
+
+def parse_blazemeter_test_link(link):
+    """
+    https://a.blazemeter.com/app/#/accounts/97961/workspaces/89846/projects/229969/tests/5823512
+
+    :param link:
+    :return:
+    """
+    if not isinstance(link, (string_types, text_type)):
+        return None
+
+    regex = r'https://a.blazemeter.com/app/#/accounts/(\d+)/workspaces/(\d+)/projects/(\d+)/tests/(\d+)(?:/\w+)?'
+    match = re.match(regex, link)
+    if match is None:
+        return None
+
+    TestParams = namedtuple('TestParams', 'account_id,workspace_id,project_id,test_id')
+    return TestParams(*[int(x) for x in match.groups()])
 
 
 class BlazeMeterUploader(Reporter, AggregatorListener, MonitoringListener, Singletone):
