+++ bzt/utils.py
@@ -1264,22 +1264,3 @@ def get_host_ips(filter_loopbacks=True):
 
 def is_url(url):
     return parse.urlparse(url).scheme in ["https", "http"]
-
-
-def parse_blazemeter_test_link(link):
-    """
-    https://a.blazemeter.com/app/#/accounts/97961/workspaces/89846/projects/229969/tests/5823512
-
-    :param link:
-    :return:
-    """
-    if not isinstance(link, (string_types, text_type)):
-        return None
-
-    regex = r'https://a.blazemeter.com/app/#/accounts/(\d+)/workspaces/(\d+)/projects/(\d+)/tests/(\d+)(?:/\w+)?'
-    match = re.match(regex, link)
-    if match is None:
-        return None
-
-    TestParams = namedtuple('TestParams', 'account_id,workspace_id,project_id,test_id')
-    return TestParams(*[int(x) for x in match.groups()])
